Allow toDate() to accept Temporal.Absolute and Temporal.DateTime
If given a Temporal.Absolute, the return value is unambiguous.

If given a Temporal.DateTime, it is interpreted as if it were in the
local time zone. This matches the behaviour of `new Date(y, m, d, ...)`.

Note that we will no longer need to pass input parameters through this
function, only return values, because we will do all calculations
internally using Temporal.